# refactor(ndt_omp): apply static analysis