--- a/include/estimate_covariance/estimate_covariance.hpp
+++ b/include/estimate_covariance/estimate_covariance.hpp
@@ -5,6 +5,10 @@
 
 #include <Eigen/Core>
 
+#include <memory>
+#include <utility>
+#include <vector>
+
 namespace pclomp
 {
 
@@ -17,17 +21,17 @@ struct ResultOfMultiNdtCovarianceEstimation
 };
 
 /** \brief Estimate functions */
-Eigen::Matrix2d estimate_xy_covariance_by_Laplace_approximation(
+Eigen::Matrix2d estimate_xy_covariance_by_laplace_approximation(
   const Eigen::Matrix<double, 6, 6> & hessian);
 ResultOfMultiNdtCovarianceEstimation estimate_xy_covariance_by_multi_ndt(
   const NdtResult & ndt_result,
-  std::shared_ptr<pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZ, pcl::PointXYZ>>
-    ndt_ptr,
+  const std::shared_ptr<
+    pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZ, pcl::PointXYZ>> & ndt_ptr,
   const std::vector<Eigen::Matrix4f> & poses_to_search);
 ResultOfMultiNdtCovarianceEstimation estimate_xy_covariance_by_multi_ndt_score(
   const NdtResult & ndt_result,
-  std::shared_ptr<pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZ, pcl::PointXYZ>>
-    ndt_ptr,
+  const std::shared_ptr<
+    pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZ, pcl::PointXYZ>> & ndt_ptr,
   const std::vector<Eigen::Matrix4f> & poses_to_search, const double temperature);
 
 /** \brief Find rotation matrix aligning covariance to principal axes