--- a/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp
+++ b/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp
@@ -194,8 +194,8 @@ MultiGridNormalDistributionsTransform<
   params_.regularization_scale_factor = 0.0f;
   params_.use_line_search = false;
 
-  double gauss_c1;
-  double gauss_c2;
+  double gauss_c1 = NAN;
+  double gauss_c2 = NAN;
 
   // Initializes the gaussian fitting parameters (eq. 6.8) [Magnusson 2009]
   gauss_c1 = 10.0 * (1 - outlier_ratio_);
@@ -213,8 +213,8 @@ void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeTra
   nr_iterations_ = 0;
   converged_ = false;
 
-  double gauss_c1;
-  double gauss_c2;
+  double gauss_c1 = NAN;
+  double gauss_c2 = NAN;
 
   // Initializes the gaussian fitting parameters (eq. 6.8) [Magnusson 2009]
   gauss_c1 = 10 * (1 - outlier_ratio_);
@@ -252,7 +252,7 @@ void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeTra
   Eigen::Matrix<double, 6, 6> hessian;
 
   double score = 0;
-  double delta_p_norm;
+  double delta_p_norm = NAN;
 
   if (regularization_pose_) {
     Eigen::Transform<float, 3, Eigen::Affine, Eigen::ColMajor> regularization_pose_transformation;
@@ -506,12 +506,12 @@ void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeAng
   Eigen::Matrix<double, 6, 1> & p, bool compute_hessian)
 {
   // Simplified math for near 0 angles
-  double cx;
-  double cy;
-  double cz;
-  double sx;
-  double sy;
-  double sz;
+  double cx = NAN;
+  double cy = NAN;
+  double cz = NAN;
+  double sx = NAN;
+  double sy = NAN;
+  double sz = NAN;
   if (fabs(p(3)) < 10e-5) {
     // p(3) = 0;
     cx = 1.0;
@@ -874,7 +874,7 @@ double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::trialVal
     // Equation 2.4.5 [Sun, Yuan 2006]
     double a_s = a_l - (a_l - a_t) / (g_l - g_t) * g_l;
 
-    double a_t_next;
+    double a_t_next = NAN;
 
     if (std::fabs(a_c - a_t) < std::fabs(a_s - a_t))
       a_t_next = a_c;