## Description

This PR applies some suggestions from the linter to code.
Checked with clang-tidy and cppcheck.

needs to be merged with https://github.com/autowarefoundation/autoware.universe/pull/8124

**Fixed**:
- use explicit cast
- add initialization [cppcoreguidelines-init-variables]
- add missing headers
- change function/method
  - use lowercase
  - add keyword to function/method which linter suggested
- use prefix `++` operator
  - for performance
- removed
  - unused var, if branch


**Not Fixed**:
- error from under /usr/*
  - currently, it can't suppressed the errors using the autoware's settings: https://github.com/orgs/autowarefoundation/discussions/4876#discussioncomment-9900497
- safe access to members of unions by boost::variant
  - speed concern
- pointer arithmetic ([cppcoreguidelines-pro-bounds-pointer-arithmetic](https://releases.llvm.org/14.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/cppcoreguidelines-pro-bounds-pointer-arithmetic.html))
- [readability-function-cognitive-complexity] from clang-tidy of function `applyFilter`
- using reinterpret_cast
- [cppcoreguidelines-macro-usage](https://releases.llvm.org/14.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/cppcoreguidelines-macro-usage.html)
  - https://github.com/tier4/ndt_omp/blob/ef654bdb1a3fe0bae6f3247b4959e230e63ba429/include/pclomp/voxel_grid_covariance_omp_impl.hpp#L420
- [cppcoreguidelines-pro-type-member-init](https://releases.llvm.org/14.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/cppcoreguidelines-pro-type-member-init.html)
  - for `thread_num_` in https://github.com/tier4/ndt_omp/blob/ef654bdb1a3fe0bae6f3247b4959e230e63ba429/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp#L105 which initialized in the method
- [openmp-use-default-none]
- [missingMemberCopy]
  - because of comment from https://github.com/tier4/ndt_omp/blob/ef654bdb1a3fe0bae6f3247b4959e230e63ba429/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp#L70
- (deprecated warning for `testGradient`, not scoped for this PR)


## Interface Changes
This PR will change the [estimate_xy_covariance_by_Laplace_approximation](https://github.com/tier4/ndt_omp/blob/ef654bdb1a3fe0bae6f3247b4959e230e63ba429/src/estimate_covariance/estimate_covariance.cpp#L7) to `estimate_xy_covariance_by_laplace_approximation` (lowercase `laplace`) which effect the [ndt_scan_matcher](https://github.com/autowarefoundation/autoware.universe/tree/main/localization/ndt_scan_matcher).


## Checks
Checked with clang-tidy (v14.0.0) and cppcheck (v2.14.1).

Use `.cppcheck_suppressions` for suppressing some notation from cppcheck.

For ignoring some notes using the autoware, autoware.universe repo's suppression files , you need to add followings:
- `CPPLINT.cfg`: `filter=-legal/copyright`
- `.cppcheck_supressions`:  `missingMemberCopy`

<details>
<summary>check_linter.sh</summary>

```
#!/bin/bash
set -eux

TARGET_DIR=$1

current_dir=$(basename $(pwd))
if [[ ! $current_dir =~ ^(autoware|pilot-auto) ]]; then
    echo "This script must be run in a directory with a prefix of autoware or pilot-auto."
    exit 1
fi

set +eux
export CPLUS_INCLUDE_PATH=/usr/include/c++/11:/usr/include/x86_64-linux-gnu/c++/11:$CPLUS_INCLUDE_PATH
set -eux

fdfind -e cpp -e hpp --full-path ${TARGET_DIR} | xargs -P $(nproc) -I{} cpplint {}
fdfind -e cpp -e hpp --full-path ${TARGET_DIR} | xargs -P $(nproc) -I{} clang-tidy -p build/ {}
```
</details>

---

Before fixing the code:
<details>
<summary>Check with check_linter.sh</summary>

```Text
$ ./check_linter.sh ndt_omp
...
Done processing ./src/universe/external/ndt_omp/src/multigrid_pclomp/multi_voxel_grid_covariance_omp.cpp
Done processing ./src/universe/external/ndt_omp/src/pclomp/gicp_omp.cpp
./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp:27:  Single-parameter constructors should be marked explicit.  [runtime/explicit] [5]
Done processing ./src/universe/external/ndt_omp/src/multigrid_pclomp/multigrid_ndt_omp.cpp
Done processing ./src/universe/external/ndt_omp/src/pclomp/ndt_omp.cpp
Done processing ./src/universe/external/ndt_omp/src/pclomp/voxel_grid_covariance_omp.cpp
Done processing ./src/universe/external/ndt_omp/apps/timer.hpp
./src/universe/external/ndt_omp/apps/util.hpp:27:  Add #include <algorithm> for sort  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/apps/util.hpp:56:  Add #include <string> for string  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/apps/util.hpp:56:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/apps/util.hpp
Total errors found: 3
./src/universe/external/ndt_omp/include/estimate_covariance/estimate_covariance.hpp:29:  Add #include <memory> for shared_ptr<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/estimate_covariance/estimate_covariance.hpp:54:  Add #include <utility> for pair<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/estimate_covariance/estimate_covariance.hpp:55:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/estimate_covariance/estimate_covariance.hpp
Total errors found: 3
./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp:84:  Add #include <map> for map<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp:85:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp:87:  Add #include <utility> for pair<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp:87:  Add #include <string> for string  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/apps/pcd_map_grid_manager.hpp
Total errors found: 5
Done processing ./src/universe/external/ndt_omp/include/pclomp/ndt_struct.hpp
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:55:  Using deprecated casting style.  Use static_cast<int>(...) instead  [readability/casting] [4]
Done processing ./src/universe/external/ndt_omp/apps/align.cpp
Done processing ./src/universe/external/ndt_omp/apps/regression_test.cpp
Done processing ./src/universe/external/ndt_omp/src/estimate_covariance/estimate_covariance.cpp
Done processing ./src/universe/external/ndt_omp/apps/check_covariance.cpp
./src/universe/external/ndt_omp/include/pclomp/voxel_grid_covariance_omp_impl.hpp:204:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/voxel_grid_covariance_omp_impl.hpp:204:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:243:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:373:  Using deprecated casting style.  Use static_cast<double>(...) instead  [readability/casting] [4]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:382:  Using deprecated casting style.  Use static_cast<double>(...) instead  [readability/casting] [4]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:383:  Using deprecated casting style.  Use static_cast<double>(...) instead  [readability/casting] [4]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:395:  Do not use namespace using-directives.  Use using-declarations instead.  [build/namespaces] [5]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:439:  Using deprecated casting style.  Use static_cast<double>(...) instead  [readability/casting] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:209:  Add #include <utility> for move  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:282:  Add #include <string> for string  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:282:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:340:  Add #include <map> for map<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:456:  Add #include <limits> for numeric_limits<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp
Total errors found: 5
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:536:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/voxel_grid_covariance_omp_impl.hpp:351:  Add #include <limits> for numeric_limits<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/pclomp/voxel_grid_covariance_omp_impl.hpp:429:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/pclomp/voxel_grid_covariance_omp_impl.hpp
Total errors found: 4
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:484:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp:492:  Add #include <utility> for pair<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/pclomp/gicp_omp_impl.hpp
Total errors found: 10
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:789:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:789:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:796:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:796:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:807:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:797:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:797:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:836:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:836:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:804:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:804:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:815:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:854:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:854:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:844:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:844:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:878:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:878:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:862:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:862:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:906:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:886:  An else should appear on the same line as the preceding }  [whitespace/newline] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:886:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:914:  If an else has a brace on one side, it should have it on both  [readability/braces] [5]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:149:  Add #include <utility> for move  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:994:  Add #include <algorithm> for max  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp:1131:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/multigrid_pclomp/multigrid_ndt_omp_impl.hpp
Total errors found: 15
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:1002:  Add #include <algorithm> for max  [build/include_what_you_use] [4]
./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp:1207:  Add #include <vector> for vector<>  [build/include_what_you_use] [4]
Done processing ./src/universe/external/ndt_omp/include/pclomp/ndt_omp_impl.hpp
```
</details>

<details>
<summary>Check with cppcheck:</summary>

```
(in .../ndt_omp)
$ cppcheck --enable=all --enable=style --enable=performance --enable=portability --enable=unusedFunction --inconclusive --check-level=exhaustive --suppressions-list=$HOME/autoware/src/universe/autoware.universe/.cppcheck_suppressions ./apps ./src/ -I include/

Checking apps/align.cpp ...
include/pclomp/gicp_omp.h:343:5: style: Struct 'OptimizationFunctorWithIndices' has a constructor with 1 argument that is not explicit. [noExplicitConstructor]
    OptimizationFunctorWithIndices(const GeneralizedIterativeClosestPoint * gicp)
    ^
include/pclomp/ndt_omp.h:186:3: performance: Function 'getFinalTransformationArray()' should return member 'transformation_array_' by const reference. [returnByReference]
  getFinalTransformationArray() const
  ^
include/multigrid_pclomp/multigrid_ndt_omp.h:232:3: performance: Function 'getFinalTransformationArray()' should return member 'transformation_array_' by const reference. [returnByReference]
  getFinalTransformationArray() const
  ^
apps/align.cpp:91:11: style: Variable 'aligned' is reassigned a value before the old one has been used. [redundantAssignment]
  aligned = align(ndt, target_cloud, source_cloud);
          ^
apps/align.cpp:85:11: note: aligned is assigned
  aligned = align(gicp_omp, target_cloud, source_cloud);
          ^
apps/align.cpp:91:11: note: aligned is overwritten
  aligned = align(ndt, target_cloud, source_cloud);
          ^
apps/align.cpp:116:13: style: Variable 'aligned' is reassigned a value before the old one has been used. [redundantAssignment]
    aligned = align(mg_ndt_omp, target_cloud, source_cloud);
            ^
apps/align.cpp:111:15: note: aligned is assigned
      aligned = align(ndt_omp, target_cloud, source_cloud);
              ^
apps/align.cpp:116:13: note: aligned is overwritten
    aligned = align(mg_ndt_omp, target_cloud, source_cloud);
            ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp.h:336:17: style: Variable 'thread_num_' is reassigned a value before the old one has been used. [redundantAssignment]
    thread_num_ = thread_num;
                ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp.h:333:19: note: thread_num_ is assigned
      thread_num_ = 1;
                  ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp.h:336:17: note: thread_num_ is overwritten
    thread_num_ = thread_num;
                ^
include/pclomp/voxel_grid_covariance_omp.h:389:88: performance: Prefer prefix ++/-- operators for non-primitive types. [postfixOperator]
    for (std::vector<int>::iterator iter = k_indices.begin(); iter != k_indices.end(); iter++) {
                                                                                       ^
include/pclomp/voxel_grid_covariance_omp.h:439:88: performance: Prefer prefix ++/-- operators for non-primitive types. [postfixOperator]
    for (std::vector<int>::iterator iter = k_indices.begin(); iter != k_indices.end(); iter++) {
                                                                                       ^
1/9 files checked 17% done
Checking apps/check_covariance.cpp ...
2/9 files checked 51% done
Checking apps/regression_test.cpp ...
apps/pcd_map_grid_manager.hpp:27:3: style: Class 'MapGridManager' has a constructor with 1 argument that is not explicit. [noExplicitConstructor]
  MapGridManager(const pcl::PointCloud<pcl::PointXYZ>::Ptr target_cloud)
  ^
3/9 files checked 69% done
Checking src/estimate_covariance/estimate_covariance.cpp ...
4/9 files checked 95% done
Checking src/multigrid_pclomp/multi_voxel_grid_covariance_omp.cpp ...
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:132:36: warning: inconclusive: Access of moved variable 'other'. [accessMoved]
  voxel_centroids_ptr_ = std::move(other.voxel_centroids_ptr_);
                                   ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:131:37: note: Calling std::move(other)
  pcl::VoxelGrid<PointT>::operator=(std::move(other));
                                    ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:132:36: note: Access of moved variable 'other'.
  voxel_centroids_ptr_ = std::move(other.voxel_centroids_ptr_);
                                   ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:133:27: warning: inconclusive: Access of moved variable 'other'. [accessMoved]
  sid_to_iid_ = std::move(other.sid_to_iid_);
                          ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:131:37: note: Calling std::move(other)
  pcl::VoxelGrid<PointT>::operator=(std::move(other));
                                    ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:133:27: note: Access of moved variable 'other'.
  sid_to_iid_ = std::move(other.sid_to_iid_);
                          ^
include/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp:134:26: warning: inconclusive: Access of moved variable 'other'. [accessMoved]
…