@a-maumau 2024-08-06
I also re-checked with updated `.cppcheck_suppressions` in autoware.universe and apply the changed at 76a9da9 .

You may need to use `--inline-suppr` to enable the inline suppression feature for checking with cppcheck.