@a-maumau 2024-08-07
To avoid the build error until merging the https://github.com/autowarefoundation/autoware.universe/pull/8124 , I added `estimate_xy_covariance_by_Laplace_approximation` for the temporal compatibility.